Fix the email rendering bug (complement 7c7ced3)
Description: When using a specific template (message_template) for the email with
Django1.8, the rendering crashes.

Cause: From Django1.8, get_template returns a backend-dependent Template
which requires a dictionary as a context for 'render'.

7c7ced3 forces the context to be a dictionary if django > 1.8. However, if
a specific email template is provided, the template created is a
django.template.Template which requires a Context object for 'render'.

Fix: This commit forces the context to be a Context object if a specific
email template has been provided.